server/requirements.txt
mcp>=1.2.0
python-telegram-bot>=21.0
google-genai>=0.2.0
fastapi
uvicorn
python-dotenv
